125799: removed the javascript automation tests
converted all page objects to typescript
converted commands to typescript
