+++ ConcernsCaseWork/ConcernsCaseWork.CypressTests/cypress/e2e/casework-regression/close-case-with-actions.cy.ts
@@ -30,7 +30,6 @@ import { EditTrustFinancialForecastPage } from "cypress/pages/caseActions/trustF
 import { ViewTrustFinancialForecastPage } from "cypress/pages/caseActions/trustFinancialForecast/viewTrustFinancialForecastPage";
 import { CloseTrustFinancialForecastPage } from "cypress/pages/caseActions/trustFinancialForecast/closeTrustFinancialForecastPage";
 import actionSummaryTable from "cypress/pages/caseActions/summary/actionSummaryTable";
-import caseActionTable from "cypress/pages/caseRows/caseActionTable";
 
 describe("Testing closing of cases when there are case actions and concerns", () =>
 {
