+++ ConcernsCaseWork/ConcernsCaseWork.CypressTests/cypress/pages/closedCasePage.ts
@@ -1,0 +1,8 @@
+class ClosedCasePage {
+    getClosedCase(caseid) {
+        return cy.get('a[href="/case/'+caseid+'/closed"]');
+       }
+}
+
+    
+    export default new ClosedCasePage();
