+++ ConcernsCaseWork/ConcernsCaseWork.CypressTests/cypress/pages/editConcernPage.ts
@@ -1,0 +1,15 @@
+class EditConcernPage {
+    closeConcern()
+    {
+        cy.getByTestId("close-concern-button").click();
+
+        return this;
+    }
+
+    confirmCloseConcern()
+    {
+        cy.getByTestId("close-concern-button").click();
+    }
+}
+
+export default new EditConcernPage();
