--- ConcernsCaseWork/ConcernsCaseWork.CypressTests/cypress/support/constants.js
@@ -1,1 +1,0 @@
-export const LogTask = "log";
